backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
decorator==5.1.1
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
django-sslserver==0.22
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
ExifRead==3.0.0
fiona==1.9.6
fonttools==4.53.0
geographiclib==2.0
geojson==3.1.0
geopandas==0.14.4
geopy==2.4.1
gpxpy==1.6.2
idna==3.7
ipython==8.26.0
jedi==0.19.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
piexif==1.1.3
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure-eval==0.2.2
py-spy==0.3.14
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyproj==3.6.1
pysnowflake==0.1.3
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scipy==1.13.1
shapely==2.0.4
six==1.16.0
snowflake==0.0.4
snowflake-id==1.0.2
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
